pyproject.toml
[project]
name = "loom"
version = "0.1.0"
description = "Local model orchestration engine — task decomposition, execution, and verification"
requires-python = ">=3.11"
dependencies = [
    # API server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.0.0",

    # HTTP client (for model APIs)
    "httpx>=0.27.0",

    # Data handling
    "pyyaml>=6.0.2",
    "pydantic>=2.10.0",

    # Terminal UI
    "textual>=1.0.0",

    # Async
    "anyio>=4.0.0",
    "aiosqlite>=0.20.0",

    # CLI
    "click>=8.1.0",

    # Office document support
    "python-docx>=1.1.0",
    "python-pptx>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.8.0",
    "pytest-cov>=5.0.0",
]
pdf = [
    "pypdf>=4.0.0",
]
mcp = [
    "mcp>=1.0.0",
]

[project.scripts]
loom = "loom.__main__:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/loom"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]

[tool.coverage.run]
source = ["loom"]

[tool.coverage.report]
show_missing = true
skip_empty = true
